app_code/requirements.txt
Flask>=2.3.3
flask-cors>=4.0.0
flask-sock>=0.7.0
boto3>=1.28.44
requests>=2.31.0
pypdf>=3.16.0
python-dotenv>=1.0.0
openai>=1.6.1
gunicorn>=21.2.0
docx2txt>=0.8
python-docx>=0.8.11
werkzeug
